requirements.txt
flask
Pillow
piexif
folium
matplotlib
